+++ Catalog.Application/Services/Implementations/CategoryService.cs
@@ -1,11 +1,12 @@
 ﻿using AutoMapper;
 using Catalog.Application.DTOs;
 using Catalog.Application.Exceptions;
+using Catalog.Application.Services.Interfaces;
 using Catalog.Domain.Entities;
 using Catalog.Domain.Interfaces;
 using Hangfire;
 
-namespace Catalog.Application.Services
+namespace Catalog.Application.Services.Implementations
 {
     public class CategoryService : ICategoryService
     {
@@ -38,22 +39,17 @@ namespace Catalog.Application.Services
 
         public async Task<Category> GetCategoryAsync(int id)
         {
-            Category category;
+            Category category = new();
             var categories = await _cacheRepository.GetDataAsync<IEnumerable<Category>>("category");
 
             if (categories != null)
             {
                 category = categories.FirstOrDefault(category => category.Id == id)!;
-
-                if (category != null)
-                {
-                    return category;
-                }
             }
 
-            category = await _unitOfWork.Categories.GetAsync(category => category.Id == id);
+            var categoryResult = category is null ? await _unitOfWork.Categories.GetAsync(category => category.Id == id) : category;
 
-            return category ?? throw new NotFoundException("Category not found");
+            return categoryResult ?? throw new NotFoundException("Category not found");
         }
 
         public async Task<Category> CreateCategoryAsync(CreateCategoryDto createCategoryDto)
@@ -93,7 +89,7 @@ namespace Catalog.Application.Services
         {
             var category = await _unitOfWork.Categories.GetAsync(category => category.Id == id)
                 ?? throw new NotFoundException("Category not found");
-            
+
             _unitOfWork.Categories.Remove(category);
             await _unitOfWork.CommitAsync();
