+++ Catalog.Application/Services/Interfaces/IReportService.cs
@@ -1,7 +1,7 @@
 ﻿using Catalog.Application.DTOs;
 using Catalog.Domain.Entities.Mongo;
 
-namespace Catalog.Application.Services
+namespace Catalog.Application.Services.Interfaces
 {
     public interface IReportService
     {
