+++ Identity.Application/Services/Implementations/TokenService.cs
@@ -1,9 +1,10 @@
-﻿using Microsoft.IdentityModel.Tokens;
+﻿using Identity.Application.Services.Interfaces;
+using Microsoft.IdentityModel.Tokens;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
 
-namespace Identity.Application.Services
+namespace Identity.Application.Services.Implementations
 {
     public class TokenService : ITokenService
     {
