+++ Identity.Web/HostingExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using Identity.Application.AutoMapper;
-using Identity.Application.Services;
+using Identity.Application.Services.Implementations;
+using Identity.Application.Services.Interfaces;
 using Identity.Web.Extensions;
 using Identity.Web.Middleware;
